Replace old ImageModalCard component with a new implementation for improved performance and preload handling
